Events: Migrate sunnypilot onroad events to its own cereal
